+++ example/lib/ui/indicator/IndicatorPage.dart
@@ -63,9 +63,19 @@ class _IndicatorPageState extends State<IndicatorPage> {
       },imgRes: "images/classical_unfollow.gif"),
       IndicatorItem(title: "QQ头部指示器",onClick: (){
         Navigator.of(context).push(MaterialPageRoute(builder: (BuildContext context) =>
-            IndicatorActivity(reverse: true,title:"QQ头部指示器" ,header: WaterDropHeader())
+            IndicatorActivity(reverse: false,title:"QQ头部指示器" ,header: WaterDropHeader())
         ));
-      },imgRes: "images/warterdrop.gif")
+      },imgRes: "images/warterdrop.gif"),
+      IndicatorItem(title: "经典Material指示器",onClick: (){
+        Navigator.of(context).push(MaterialPageRoute(builder: (BuildContext context) =>
+            IndicatorActivity(reverse: true,title:"官方Material指示器" ,header: MaterialClassicHeader())
+        ));
+      },imgRes: "images/material_classic.gif"),
+      IndicatorItem(title: "水滴坠落Material指示器",onClick: (){
+        Navigator.of(context).push(MaterialPageRoute(builder: (BuildContext context) =>
+            IndicatorActivity(reverse: true,title:"水滴坠落Material指示器" ,header: WaterDropMaterialHeader())
+        ));
+      },imgRes: "images/material_waterdrop.gif"),
     ];
     super.didChangeDependencies();
   }