+++ example/lib/ui/test/Example2.dart
@@ -93,21 +93,17 @@ class _Example2State extends State<Example2> with TickerProviderStateMixin {
     return SmartRefresher(
       enablePullDown: true,
       enablePullUp: true,
-      header: WaterDropHeader(),
+      header: WaterDropMaterialHeader(),
       controller: _controller,
       onRefresh: _onRefresh,
       onLoading: _onLoading,
       onOffsetChange: _onOffsetCallback,
-      child: new StaggeredGridView.countBuilder(
+      child: GridView.builder(
         key: PageStorageKey("w"),
-        crossAxisCount: 4,
         controller: _scrollController,
+        gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(crossAxisCount: 2),
         itemCount: data.length,
-        itemBuilder: (BuildContext context, int index) => Item(url:data[index]),
-        staggeredTileBuilder: (int index) =>
-            new StaggeredTile.count(2, index.isEven ? 2 : 1),
-        mainAxisSpacing: 4.0,
-        crossAxisSpacing: 4.0,
+        itemBuilder: buildImage,
       ),
     );
   }